Some noise work

+++ VolatileHordes/Settings/User/Control/LuckyPlayerRetargetSettings.cs
@@ -2,6 +2,6 @@
 {
     public class LuckyPlayerRetargetSettings
     {
-        public double ChancePerMinute { get; set; } = 0.02;
+        public float ChancePerMinute { get; set; } = 0.02f;
     }
 }
